requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
boto3==1.24.82
botocore==1.27.82
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Cython==0.29.32
Django==4.1.1
django-crispy-forms==1.14.0
django-mysql==4.7.1
django-paypal==2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-ses==3.1.2
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
fontawesomefree==6.2.0
idna==3.4
jmespath==1.0.1
paypalrestsdk==1.13.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
pyngrok==5.1.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
selectolax==0.3.11
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
textdistance==4.5.0
tzdata==2022.2
urllib3==1.26.12
